requirements.txt
weave~=0.51.8
openai~=1.40.6
python-dotenv~=1.0.1